+++ src/main/java/org/folio/linked/data/integration/ResourceModificationEventListener.java
@@ -1,7 +1,7 @@
 package org.folio.linked.data.integration;
 
 import static org.folio.linked.data.util.Constants.STANDALONE_PROFILE;
-import static org.folio.linked.data.util.ResourceUtils.getTypes;
+import static org.folio.linked.data.util.ResourceUtils.getTypeUris;
 
 import java.util.Collection;
 import lombok.RequiredArgsConstructor;
@@ -39,7 +39,7 @@ public class ResourceModificationEventListener {
   public void afterCreate(ResourceCreatedEvent resourceCreatedEvent) {
     var resource = resourceCreatedEvent.resource();
     log.debug("ResourceCreatedEvent received [{}]", resourceCreatedEvent);
-    log.info("Resource with id {} and types {} was created", resource.getId(), getTypes(resource));
+    log.info("Resource with id {} and types {} was created", resource.getId(), getTypeUris(resource));
     createMessageSenders.forEach(sender -> sender.produce(resource));
   }
 
@@ -47,7 +47,7 @@ public class ResourceModificationEventListener {
   public void afterUpdate(ResourceUpdatedEvent resourceUpdatedEvent) {
     var resource = resourceUpdatedEvent.resource();
     log.debug("ResourceUpdatedEvent received [{}]", resourceUpdatedEvent);
-    log.info("Resource with id {} and types {} was updated", resource.getId(), getTypes(resource));
+    log.info("Resource with id {} and types {} was updated", resource.getId(), getTypeUris(resource));
     updateMessageSenders.forEach(sender -> sender.produce(resource));
   }
 
@@ -57,7 +57,7 @@ public class ResourceModificationEventListener {
     var currentResourceId = resourceReplacedEvent.currentResourceId();
     log.debug("ResourceReplacedEvent received [{}]", resourceReplacedEvent);
     log.info("Resource with id {} and types {} was replaced by resource with id {}",
-      previous.getId(), getTypes(previous), currentResourceId);
+      previous.getId(), getTypeUris(previous), currentResourceId);
     resourceRepository.findById(currentResourceId)
       .ifPresent(
         resource -> replaceMessageSenders.forEach(sender -> sender.produce(previous, resource))
@@ -68,7 +68,7 @@ public class ResourceModificationEventListener {
   public void afterDelete(ResourceDeletedEvent resourceDeletedEvent) {
     var resource = resourceDeletedEvent.resource();
     log.debug("ResourceDeletedEvent received [{}]", resourceDeletedEvent);
-    log.info("Resource with id {} and types {} was deleted", resource.getId(), getTypes(resource));
+    log.info("Resource with id {} and types {} was deleted", resource.getId(), getTypeUris(resource));
     deleteMessageSenders.forEach(sender -> sender.produce(resource));
   }