+++ src/main/java/org/folio/linked/data/model/entity/Profile.java
@@ -7,14 +7,12 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.Data;
 import lombok.experimental.Accessors;
 import org.hibernate.annotations.Type;
 
+@Data
 @Entity
-@Getter
-@Setter
 @Table(name = "profiles")
 @Accessors(chain = true)
 public class Profile {